requirements.txt
django
pillow
django-ckeditor
django-widget-tweaks